Keep pipelines BulkRequestPreprocessor into the IngestService for simplicity, wire Node and security
